src/kuzu/requirements.txt
kuzu==0.7.0
polars==1.16.0
pyarrow==18.1.0
python-dotenv==1.0.1
ell-ai[all]==0.0.15